feat: initialize api server
