+++ app.js
@@ -6,7 +6,7 @@ const session = require('express-session')
 const passport = require('./config/passport')
 const handlebarsHelpers = require('./helpers/handlebars-helpers')
 const { getUser } = require('./helpers/auth-helpers')
-const routes = require('./routes')
+const { pages } = require('./routes')
 if (process.env.NODE_ENV !== 'production') {
   require('dotenv').config()
 }
@@ -32,7 +32,7 @@ app.use((req, res, next) => {
   res.locals.user = getUser(req)
   next()
 })
-app.use(routes)
+app.use(pages)
 
 app.listen(port, () => {
   console.info(`Example app listening on port ${port}!`)
