+++ controllers/pages/user-controller.js
@@ -1,6 +1,6 @@
 const bcrypt = require('bcryptjs') // 載入 bcrypt
-const { imgurFileHandler } = require('../helpers/file-helpers')
-const { User, Restaurant, Comment, Favorite, Like, Followship } = require('../models')
+const { imgurFileHandler } = require('../../helpers/file-helpers')
+const { User, Restaurant, Comment, Favorite, Like, Followship } = require('../../models')
 const userController = {
   // 負責 render 註冊的頁面
   signUpPage: (req, res) => {
